requirements.txt
watchdog
yara-python
logging
